ffmpeg_next::format::stream::disposition

结构体 Disposition

pub struct Disposition(/* 私有字段 */);

实现

impl Disposition

pub const DEFAULT: Self

pub const DUB: Self

pub const ORIGINAL: Self

pub const COMMENT: Self

pub const LYRICS: Self

pub const KARAOKE: Self

pub const FORCED: Self

pub const HEARING_IMPAIRED: Self

pub const VISUAL_IMPAIRED: Self

pub const CLEAN_EFFECTS: Self

pub const ATTACHED_PIC: Self

pub const CAPTIONS: Self

pub const DESCRIPTIONS: Self

pub const METADATA: Self

pub const MULTILAYER: Self

impl Disposition

pub const fn empty() -> Self

获取一个所有位均为 0 的 flags 值。

pub const fn all() -> Self

获取一个所有已知位均为 1 的 flags 值。

pub const fn bits(&self) -> c_int

获取底层的位值。返回值正是该 flags 值中已设置的位。

pub const fn from_bits(bits: c_int) -> Option<Self>

从位值转换。若设置了任何未知位，则该方法将返回 None。

pub const fn from_bits_truncate(bits: c_int) -> Self

从位值转换，未知的位会被清除。

pub const fn from_bits_retain(bits: c_int) -> Self

按位值原样转换（保留所有位）。

pub fn from_name(name: &str) -> Option<Self>

获取一个 flags 值，其位为具有给定名称的 flag 所对应的位。若 name 为空或不与任何已命名的 flag 对应，则该方法将返回 None。

pub const fn is_empty(&self) -> bool

判断 self 中的所有位是否都未设置。

pub const fn is_all(&self) -> bool

判断该 flags 值的所有已知位是否都已设置。

pub const fn intersects(&self, other: Self) -> bool

判断 other 中是否至少有一位也在 self 中被设置。

pub const fn contains(&self, other: Self) -> bool

判断 other 中所有已设置的位是否都在 self 中被设置（即 self ⊇ other）。

pub fn insert(&mut self, other: Self)

self 与 other 的按位或（|）。

pub fn remove(&mut self, other: Self)

self 与 other 的补的交集（&!）。当 other 含有未知位时，该方法不等价于 self & !other。remove 不会截断 other，而 ! 运算符会。

pub fn toggle(&mut self, other: Self)

self 与 other 的按位异或（^）。

pub fn set(&mut self, other: Self, value: bool)

当 value 为 true 时调用 insert，当 value 为 false 时调用 remove。

pub const fn intersection(self, other: Self) -> Self

self 与 other 的按位与（&）。

pub const fn union(self, other: Self) -> Self

self 与 other 的按位或（|）。

pub const fn difference(self, other: Self) -> Self

self 与 other 的补的交集（&!）。当 other 含有未知位时，该方法不等价于 self & !other。该方法不会截断 other，而 ! 运算符会。

pub const fn symmetric_difference(self, other: Self) -> Self

self 与 other 的按位异或（^）。

pub const fn complement(self) -> Self

对 self 的位按位取反（!），并截断结果。

impl Disposition

pub const fn iter(&self) -> Iter<Disposition>

Yield a set of contained flags values.

每个产出的 flag 值对应一个已定义的具名 flag。任何未知位将作为最终的flags 值一起产出。

pub const fn iter_names(&self) -> IterNames<Disposition>

逐个产出该 flags 值中所包含的已命名 flag 集合。

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.